Fix Boxes
Drawing the Boxes one at a time.
Showcases first the outer box, then the middle box, and finally the inner box.
Utilizes a sleep call to help the user visualize the steps better.
This is to ensure students understand how each box is drawn on top of each other. This is why if the inner box were to have been called first it would have been covered up.